exynos4: hwc: draw contiguous layers directly to FIMD
If a layer is contiguously allocated in gralloc (i.e. in ION or FIMC),
it has a non-zero paddr, and as long as it doesn't require FIM[CG] for some
other reason (scaling, rotating) FIMD can render it directly.

This seems to be the intended use of gsc_map_t::NONE.

Also update hwc_dump to show some relevant information about hw win to
layer mapping, window mode, and paddr of the memory.

Change-Id: I4c86655131cc859ee789fb37fa30a352376cf8cf